% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_set.R
\name{count_set}
\alias{count_set}
\title{Generate a count_set summerized experiment from NanoString data}
\usage{
count_set(
  rccexp_dir = NULL,
  count_data = NULL,
  group = NULL,
  batch = NULL,
  pair = NULL,
  samp_id = NULL,
  output_log = NULL,
  count_se = NULL,
  low_reps = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{rccexp_dir}{Full path to the csv file generated by nSolver RCC
Collector Tool Format Export. Default = NULL}

\item{count_data}{a matrix of count data; genes as rows, samples as columns.
Contains 3 columns called "Code_Class", "Gene_Name" & "Accession". Code_class
contains probe annotations "Endogenous", "Positive" and "Negative". Default = NULL}

\item{group}{Biological groups for each lane of counts defined by a factor
or vector. Default = NULL}

\item{batch}{Batch information such as nanostring chip ID, defined by a
factor or vector of integers. Default = NULL}

\item{pair}{Sample pairing defined by a factor or vector. Default = NULL}

\item{samp_id}{Sample IDs defined by a factor or vector. This is used to
build the replicate matrix. Default = NULL}

\item{output_log}{File path to directory where the summarized experiment
will be saved}

\item{count_se}{File path to an already generated summarized experiment}

\item{low_reps}{Force Ccount_set to build if a group contains less than
2 replicates. Default = FALSE}

\item{verbose}{Verbosity ON/OFF. Default = TRUE}
}
\value{
A summarized experiment
}
\description{
Generates a summerized experiment from NanoString count
data in an nSolver RCC Collector Tool Format or matrix and (optional)
sample annotations including group, batch, and sample pairing.
}
\examples{

# biological groups
rnf5_group <- c(rep("WT", 5), rep("KO", 5))

# sample ids
rnf5_sampleid <- c("GSM3638131", "GSM3638132", "GSM3638133", "GSM3638134",
                  "GSM3638135", "GSM3638136", "GSM3638137", "GSM3638138",
                  "GSM3638139", "GSM3638140")

# build count_set
rnf5_count_set <- count_set(count_data = Rnf5,
                            group = rnf5_group,
                            samp_id = rnf5_sampleid)

}
