% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{hclust_plot_wrap}
\alias{hclust_plot_wrap}
\title{hclust plot wrapper}
\usage{
hclust_plot_wrap(
  count_set = NULL,
  norm_method = "housekeeping_scaled",
  title = NA,
  legend = TRUE,
  colors = NA
)
}
\arguments{
\item{count_set}{A count_set of mRNA or miRNA counts generated using count_set}

\item{norm_method}{Name of the count_set assay to be plotted. Options are
"counts", "background_corrected","positive_control_scaled",
"housekeeping_scaled", "geNorm", "all_endogenous_scaled", "loess", "vsn", "quantile", "ruvIII".
Default = "housekeeping_scaled"}

\item{title}{Plot title}

\item{legend}{Option for plot legend. Default = TRUE}

\item{colors}{Option to define colour scheme. Default = NA}
}
\value{
hclust plots
}
\description{
hclust plot wrapper
}
