% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_rank_dunn.R
\name{norm_rank}
\alias{norm_rank}
\title{Ranks NanoString normalizations using cluster validity indices}
\usage{
norm_rank(count_set = count_set, batch_aware = FALSE)
}
\arguments{
\item{count_set}{a normalised, count_set summarized experiment.
Default = NULL}

\item{batch_aware}{TRUE or FALSE are rankings to be batch aware?
Default = FALSE.}
}
\value{
plot and table of ranked normalizations
}
\description{
ranks NanoString normalisation by Dunn Index on euclidian
and correlation distances.
}
\examples{
# biological groups
rnf5_group <- c(rep("WT", 5), rep("KO", 5))

# sample ids
rnf5_sampleid <- c("GSM3638131", "GSM3638132", "GSM3638133", "GSM3638134",
                  "GSM3638135", "GSM3638136", "GSM3638137", "GSM3638138",
                  "GSM3638139", "GSM3638140")

# build count_set
rnf5_count_set <- count_set(count_data = Rnf5,
                            group = rnf5_group,
                            samp_id = rnf5_sampleid)
# normalize
rnf5_count_set_norm <- multi_norm(count_set = rnf5_count_set,
                                  positive_control_scaling = TRUE,
                                  background_correct = "mean2sd")
                                  #plot_dir = "~/Dropbox/git/NanoStringClustR/plot_test/")
# rank normalizations
rnf5_eval <- norm_rank(rnf5_count_set_norm)

}
\references{
Lukasz Nieweglowski (2013). clv: Cluster Validation Techniques. R package version 0.3-2.1.
https://CRAN.R-project.org/package=clv
}
