% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rnf5}
\alias{Rnf5}
\title{Raw Rnf5 NanoString dataset}
\format{raw counts exported from nSolver in RCC exporter tool format}
\source{
NCBI GEO GSE127753
}
\usage{
Rnf5
}
\description{
A dataset containing raw counts for the Rnf5 dataset.
}
\references{
Li Y, Tinoco R, Elmen L, et al. Gut microbiota dependent anti-tumor
 immunity restricts melanoma growth in Rnf5(-/-) mice. Nat Commun. 2019 Apr 2;10(1):1492.
 doi: 10.1038/s41467-019-09525-y. PubMed PMID: 30940817; PubMed Central PMCID: PMCPMC6445090.
}
\keyword{datasets}
