% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_rank.R
\name{norm_rank}
\alias{norm_rank}
\title{Ranks NanoString normalizations using cluster validity indices}
\usage{
norm_rank(count_set = count_set)
}
\arguments{
\item{count_set}{a normalised, count_set summarized experiment
generated by count_set.Default = NULL}
}
\value{
ranked normalizations
}
\description{
ranks NanoString normalisation by 1) clustering of
groups and; 2) variation in relative log expression.
Clustering is assessed by a Generalised Dunn Index using the
average linkage and average diamter for inter and intracluster
distances, respectively. Overall variation is assessed by Relative
Log Expression.
}
\examples{
# biological groups
rnf5_group <- c(rep("WT", 5), rep("KO", 5))

# sample ids
rnf5_sampleid <- c("GSM3638131", "GSM3638132", "GSM3638133", "GSM3638134",
                  "GSM3638135", "GSM3638136", "GSM3638137", "GSM3638138",
                  "GSM3638139", "GSM3638140")

# build count_set
rnf5_count_set <- count_set(count_data = Rnf5,
                            group = rnf5_group,
                            samp_id = rnf5_sampleid)
# normalize
rnf5_count_set_norm <- multi_norm(count_set = rnf5_count_set,
                                  positive_control_scaling = TRUE,
                                  background_correct = "mean2sd")
                                  #plot_dir = "~/Dropbox/git/NanoStringClustR/plot_test/")
# rank normalizations
rnf5_eval <- norm_rank(rnf5_count_set_norm)

}
\references{
Lukasz Nieweglowski (2013). clv: Cluster Validation Techniques.
R package version 0.3-2.1. https://CRAN.R-project.org/package=clv
}
